+++ etna/loggers/__init__.py
@@ -16,7 +16,7 @@ from etna import SETTINGS
 from etna.loggers.base import _Logger
 from etna.loggers.console_logger import ConsoleLogger
 
-if SETTINGS.wandb_required and SETTINGS.torch_required:
+if SETTINGS.wandb_required:
     from etna.loggers.wandb_logger import WandbLogger
 
 tslogger = _Logger()